+++ app/muveletek/mTorles.php
@@ -1,7 +1,11 @@
 <?php
-require("kapcsolat.php");
-$torlendo = mysqli_real_escape_string($dbconn, $_GET['eftorlendo']);
+if(!defined('eleres')){
+    header("Location: ../hiba.html");
+} else{
+    require("kapcsolat.php");
+    $torlendo = mysqli_real_escape_string($dbconn, $_GET['eftorlendo']);
 
-//Törlés
-mysqli_query($dbconn, "DELETE FROM edzoklienskapcs WHERE kapcs_id = {$torlendo}");
+    //Törlés
+    mysqli_query($dbconn, "DELETE FROM edzoklienskapcs WHERE kapcs_id = {$torlendo}");
+}
 ?>
